GH-121 Reorganize storage module into directory with separate files
Split the flat src/storage.rs into a src/storage/ module directory and
move GitNodeStorage from src/git/storage.rs into src/storage/git.rs,
grouping all NodeStorage implementations under a unified module.

This is Phase 1 of separating git storage backend from versioning logic.
